tasks: publish workflow_uuid to MQ
