+++ reana_workflow_engine_yadage/tasks.py
@@ -26,22 +26,22 @@ import json
 import logging
 import os
 
-import pika
 import zmq
 from yadage.steering_api import steering_ctx
 from yadage.utils import setupbackend_fromstring
 
 from . import celery_zeromq
 from .celeryapp import app
-from .config import (BROKER_PASS, BROKER_PORT, BROKER_URL, BROKER_USER,
-                     CODE_DIRECTORY_RELATIVE_PATH,
+from .config import (CODE_DIRECTORY_RELATIVE_PATH,
                      INPUTS_DIRECTORY_RELATIVE_PATH,
                      LOGS_DIRECTORY_RELATIVE_PATH,
                      OUTPUTS_DIRECTORY_RELATIVE_PATH, SHARED_VOLUME_PATH,
                      YADAGE_INPUTS_DIRECTORY_RELATIVE_PATH)
+from .utils import publish_workflow_status
 from .zeromq_tracker import ZeroMQTracker
 
 log = logging.getLogger(__name__)
+
 known_dirs = [
     CODE_DIRECTORY_RELATIVE_PATH,
     INPUTS_DIRECTORY_RELATIVE_PATH,
@@ -49,35 +49,6 @@ known_dirs = [
     OUTPUTS_DIRECTORY_RELATIVE_PATH,
     YADAGE_INPUTS_DIRECTORY_RELATIVE_PATH,
 ]
-
-
-def publish_workflow_status(workflow_uuid, status, message=None):
-    """Update database workflow status.
-
-    :param workflow_uuid: UUID which represents the workflow.
-    :param status: String that represents the analysis status.
-    :param status_message: String that represents the message related with the
-       status, if there is any.
-    """
-    broker_credentials = pika.PlainCredentials(BROKER_USER,
-                                               BROKER_PASS)
-    connection = pika.BlockingConnection(
-        pika.ConnectionParameters(BROKER_URL,
-                                  BROKER_PORT,
-                                  '/',
-                                  broker_credentials))
-    channel = connection.channel()
-    channel.queue_declare(queue='jobs-status')
-    log.info('Publishing Workflow: {0} Status: {1}'.format(workflow_uuid,
-                                                           status))
-    channel.basic_publish(exchange='',
-                          routing_key='jobs-status',
-                          body=json.dumps({"workflow_uuid": workflow_uuid,
-                                           "status": status,
-                                           "message": message}),
-                          properties=pika.BasicProperties(
-                              delivery_mode=2,  # msg persistent
-                          ))
 
 
 @app.task(name='tasks.run_yadage_workflow', ignore_result=True)
@@ -88,7 +59,7 @@ def run_yadage_workflow(workflow_uuid, workflow_workspace,
 
     workflow_workspace = '{0}/{1}'.format(SHARED_VOLUME_PATH,
                                           workflow_workspace)
-
+    app.conf.update(WORKFLOW_UUID=workflow_uuid)
     zmqctx = celery_zeromq.get_context()
     socket = zmqctx.socket(zmq.PUB)
     socket.connect(os.environ['ZMQ_PROXY_CONNECT'])
